requirements.txt
# Core ML/AI Libraries
torch>=2.0.0
transformers>=4.35.0
diffusers>=0.24.0
accelerate>=0.25.0
safetensors>=0.4.0
bitsandbytes>=0.41.0

# Image Processing
Pillow>=10.0.0

# PDF Generation
reportlab>=4.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0

# Web Interface
gradio>=4.0.0

# Optional: Enhanced features
# ebooklib>=0.18  # For EPUB generation
# xformers>=0.0.22  # For faster image generation
